+++ src/main/java/com/medco/HealthConnectProvider/ui/request/auth/password/contract/ContractRequest.java
@@ -42,7 +42,7 @@ public class ContractRequest {
         private String itemUuid;
         private String serviceName;
         private String itemType;
-        private BigDecimal negotiatedPrice;
+        private Double negotiatedPrice;
         private String packageCategory;
     }
 }
